# docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dataspace Assistant",
    "version": "0.1.0",
    "description": "Typescript and Express LangChain Chatbot"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
    "/v1/chat": {
      "post": {
        "summary": "chat",
        "tags": [
          "Chat"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "result",
            "description": "optional query params to retrieve the bests result in the response",
            "in": "query",
            "required": false,
            "type": "boolean"
          },
          {
            "name": "type",
            "description": "Mandatory query params to select the type of chatbot used",
            "in": "query",
            "required": true,
            "type": "string",
            "example": "rag or agent"
          },
          {
            "name": "user",
            "description": "optional query params to select the user prompt used by the chatbot",
            "in": "query",
            "required": false,
            "type": "number",
            "example": 0
          },
          {
            "name": "system",
            "description": "optional query params to select the system prompt used by the chatbot",
            "in": "query",
            "required": false,
            "type": "number",
            "example": 0
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "description": "message for LLM",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
    "/v1/chat/{threadId}": {
      "post": {
        "summary": "chat",
        "tags": [
          "Chat"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "threadId",
            "description": "use thread.",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "message": {
                    "description": "message for LLM",
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    },
    "/v1/embedding": {
      "post": {
        "summary": "embed",
        "tags": [
          "Embedding"
        ],
        "security": [
          {
            "adminKey": []
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          }
        }
      }
    }
  },
  "components": {
    "securitySchemes": {
      "adminKey": {
        "type": "apiKey",
        "in": "header",
        "name": "x-assistant-admin-key"
      }
    }
  },
  "tags": [
    {
      "name": "Embedding",
      "description": "Embedding Routes"
    },
    {
      "name": "Chat",
      "description": "Chat Routes"
    }
  ]
}